backend/requirements.txt
amqp==5.0.3
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.5
django-admin-rangefilter==0.6.4
django-celery-results==2.0.1
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.0.4
importlib-metadata==3.4.0
kombu==5.0.2
Markdown==3.3.3
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
PyJWT==2.0.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
